Implementando labresults

+++ public/patients.php
@@ -53,8 +53,7 @@
             addPatient($conn);
         }
         else if(strcmp($operation,"DELETE_PTT"){
-            @$query = pg_query($conn, "DELETE FROM public.\"patients\" WHERE
-                                      patientid ='".$patientID."' AND userid = '".$_SESSION['id']."'");
+            @$query = pg_query($conn, "DELETE FROM public.\"patients\" WHERE patientid ='".$patientID."' AND userid = '".$_SESSION['id']."'");
         }
 
         //Returns to the original page
